LayoutTests/imported/w3c/web-platform-tests/html/canvas/offscreen/text/2d.text.drawing.style.letterSpacing.measure.worker-expected.txt


FAIL Testing letter spacing with different length units assert_approx_equals: letter spacing doesn't work. expected 9.853906250000001 +/- 0.2 but got 18.50390625
